retain the old grade value when editting
old behavior: user has a numeric grade of 85, he wants to double click it to increase it to 87 by simply clicking the small up arrow twice.
but when he clicks it, it starts counting from 0.

current behavior: when the user clicks the cell to edit it, the current value is displayed instead of zero, allowing the user to easily make small changes with the up/down arrows